test: keep CLI seams internal
